Moving char_data out of merc.h
